Fixes to laplacian kernel (don't treat like rbf), other tests fixed.

+++ falkon/tests/test_falkon.py
@@ -40,7 +40,7 @@ class TestFalkon:
         kernel = kernels.GaussianKernel(2.0)
 
         def error_fn(t, p):
-            return 100 * torch.sum(t * p <= 0) / t.shape[0], "c-err"
+            return 100 * torch.sum(t * p <= 0).to(torch.float32) / t.shape[0], "c-err"
 
         opt = FalkonOptions(use_cpu=True, no_keops=True, debug=True)
